Productsタブ用プラグインを追加。 ただし、動作しない。

+++ shared/Dtos.fs
@@ -2,18 +2,18 @@ namespace App.Shared
 
 /// ユーザーに関するDTO
 type UserDto =
-    { Id: int64
+    { Id: int
       Username: string
       Email: string }
 
 /// 製品に関するDTO
 type ProductDto =
-    { Id: int64
+    { Id: int
       Name: string
       Description: string option
       Category: string option
       Price: double
-      Stock: int64
+      Stock: int
       SKU: string
       IsActive: bool
       CreatedAt: string
